// client/src/app/api/models/music.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Playlist } from '../models/playlist';
import { User } from '../models/user';
export interface Music {
  createdAt: string;
  duration: number;
  id: string;
  lyrics: string;
  objectName: string;
  playlists: Array<Playlist>;
  prompt: string;
  publicUrl: string;
  status: string;
  style: string;
  sunoId: string;
  title: string;
  user: User;
  userId: string;
}